+++ packages/builders/radix-icons-builder/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @steeze-ui/radix-icons
 
+## 1.1.0
+
+### Minor Changes
+
+- improve typings
+- bundle sources in index.ts
+- shrinken package size
+
 ## 1.0.4
 
 ### Patch Changes